Move walker into a separate file
