+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "outDir":  "./dist",
   "compilerOptions": {
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
@@ -25,5 +26,8 @@
   "exclude": [
     "node_modules",
     "dist"
+  ],
+  "include": [
+    "src"
   ]
 }
